evaluation_requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
scikit-image>=0.18.0
tqdm>=4.50.0
torch>=1.7.0
torchvision>=0.8.0
lpips>=0.1.4
pandas>=1.2.0
matplotlib>=3.3.0
Pillow>=8.0.0 